+++ library/config/dataclasses/sdxl_finetune.py
@@ -6,9 +6,7 @@ from .model import ModelConfig
 from .sdxl import SDXLConfig
 from .performance import PerformanceConfig
 from .loss import LossConfig
-from .regularization import RegularizationConfig
 from .timestep import TimestepConfig
-from .masked_loss import MaskedLossConfig
 from .buckets import BucketsConfig
 from .output import OutputConfig
 
@@ -27,7 +25,5 @@ class SDXLFineTuneConfig:
     output: OutputConfig = field(default_factory=OutputConfig)
     performance: PerformanceConfig = field(default_factory=PerformanceConfig)
     loss: LossConfig = field(default_factory=LossConfig)
-    regularization: RegularizationConfig = field(default_factory=RegularizationConfig)
     timestep: TimestepConfig = field(default_factory=TimestepConfig)
-    masked_loss: MaskedLossConfig = field(default_factory=MaskedLossConfig)
     buckets: BucketsConfig = field(default_factory=BucketsConfig)